airflow/requirements.txt
apache-airflow-providers-google
apache-airflow-providers-apache-spark
pyarrow
kaggle
findspark